move Nav component to shared
