+++ src/components/Home/index.ts
@@ -2,4 +2,7 @@ export * from "./HomeHeader/HomeHeader";
 export * from "./HomeSections/1-categories/Categories";
 export * from "./HomeSections/2-projects/Projects";
 export * from "./HomeSections/3-partners/Partners";
+
+export * from "../shared/Nav/Nav";
 export * from "../shared/AppFooter/AppFooter";
+export * from "../shared/AllowCookies/AllowCookies";
