+++ src/components/Pages/index.ts
@@ -3,5 +3,5 @@ export * from "./OurWork/OurWork";
 export * from "./Thoughts/Thoughts";
 export * from "./WeAre/WeAre";
 
-export * from "../Nav/Nav";
+export * from "../shared/Nav/Nav";
 export * from "../shared/AppFooter/AppFooter";
