+++ src/components/shared/AppFooter/CompanyInfo/CompanyInfoSections/ChazzOffices.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 
 import offices from "../../../../../content/offices.json";
-
 import { Office } from "../../../../../interfaces/interfaces";
 
 import { LinkedEmail, LinkedMap, LinkedPhone } from "./footerLinks/index";
