+++ src/components/shared/Nav/ChazzLogo/ChazzLogo.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { Link } from "react-router-dom";
 
-import { ReactComponent as Logo } from "../../../assets/Chazz_Logo.svg";
+import { ReactComponent as Logo } from "../../../../assets/Chazz_Logo.svg";
 
 import "../Nav.styles.scss";
