+++ src/components/shared/Nav/Nav.styles.scss
@@ -1,4 +1,4 @@
-@import "../../styles.scss";
+@import "../../../styles.scss";
 
 // MOBILE FIRST
